Removed resizing of hashmaps
