+++ shared/src/question-generators/Hashing/Hashing.test.ts
@@ -282,54 +282,4 @@ describe("Hashing with Linear Probing", () => {
 |  |  |  |  |  |  |  |  |  |  |  |`,
     )
   })
-
-  test("Aufgabe 8.2 Hashing II (own single hash function) and resizing", () => {
-    /*
-    Gegeben ist eine leere Hashtabelle H der Größe m = 2 mit den Hashfunktionen hi(x) = (13x + i) mod m.
-    Fügen sie die folgende Sequenz in der gegebenen Reihenfolge in die Hashtabelle ein und vergrößern Sie
-    die Hashtabelle, wann immer der Auslastungsfaktor λ wenigstens 12 erreicht. Stellen Sie die Hashtabelle
-    vor und nach jeder Reorganisation sowie nachdem alle Schlüssel eingefügt wurden dar. Immer wenn Sie
-    die Hashtabelle vergrößern müssen wählen Sie als neue Größe die kleinste Primzahl, sodass sich die Größe
-    wenigstens verdoppelt. Fügen Sie Schlüssel, die bereits in der Hashtabelle waren, in der Reihenfolge in
-    die größere Tabelle ein, in der sie in der kleineren Tabelle standen (vom kleinen zum großen Index).
-                              90, 35, 88, 28, 91, 44, 26, 53, 16, 14, 59
-     */
-
-    const hashFunction: HashFunction = (x: number, size: number): number => {
-      return (13 * x) % size
-    }
-
-    const hashMap = new MapLinProbing({ size: 2, hashFunction, resize: true })
-
-    hashMap.insert(90, "")
-
-    expect(hashMap.toString()).toEqual(`| 0|1|2|3|4|
-|90| | | | |
-|  | | | | |`)
-
-    hashMap.insert(35, "")
-    hashMap.insert(88, "")
-
-    expect(hashMap.toString()).toEqual(`| 0|1|2|3| 4| 5|6|7|8|9|10|
-|88| | | |90|35| | | | |  |
-|  | | | |  |  | | | | |  |`)
-
-    hashMap.insert(28, "")
-    hashMap.insert(91, "")
-    hashMap.insert(44, "")
-
-    expect(hashMap.toString()).toEqual(`|0|1|2|3|4|5|6|7|8|9|10|11|12|13|14|15|16|17|18|19|20|21|22|
-| | | | | | | | | | |91|  |  |  |  |  |  |88|35|28|44|90|  |
-| | | | | | | | | | |  |  |  |  |  |  |  |  |  |  |  |  |  |`)
-
-    hashMap.insert(26, "")
-    hashMap.insert(53, "")
-    hashMap.insert(16, "")
-    hashMap.insert(14, "")
-    hashMap.insert(59, "")
-
-    expect(hashMap.toString()).toEqual(`| 0| 1|2|3|4|5|6|7| 8|9|10|11|12|13|14|15|16|17|18|19|20|21|22|
-|14|16| | | | | | |59| |91|  |  |  |  |  |26|88|35|28|44|90|53|
-|  |  | | | | | | |  | |  |  |  |  |  |  |  |  |  |  |  |  |  |`)
-  })
 })
