+++ shared/src/question-generators/Hashing/Hashing.ts
@@ -7,12 +7,10 @@ import {
 } from "@shared/api/QuestionGenerator.ts"
 import { serializeGeneratorCall } from "@shared/api/QuestionRouter.ts"
 import { generateHashFunction } from "@shared/question-generators/Hashing/Functions.ts"
-import { MapLinked } from "@shared/question-generators/Hashing/MapLinked.ts"
 import {
-  DoubleHashFunction,
-  HashFunction,
-  MapLinProbing,
-} from "@shared/question-generators/Hashing/MapLinProbing.ts"
+  createTableForValues,
+  generateOperationsHashMap,
+} from "@shared/question-generators/Hashing/utils.ts"
 import Random from "@shared/utils/random.ts"
 import { t, tFunctional, Translations } from "@shared/utils/translations.ts"
 
@@ -20,33 +18,70 @@ const translations: Translations = {
   en: {
     name: "Hashing",
     description: "Correctly insert value into a hash map",
-    resize: "resizable",
-    noResize: "non-resizable",
     insert: "We **insert** the following keys: {{0}}",
     delete: "and then **delete** the following keys: {{0}}",
-    resizeFactor:
-      "Whenever the design factor is at least $\\frac{1}{2}$ we increase the size to the next prime number, but the size at least doubles. If the design factor is at most $\\frac{1}{8}$ we decrease the size to the next prime number, but the size at least quarters.",
-    textLinear: `Consider having a empty {{0}} hash table (of size \${{1}}$) implemented with  "Linear Probing". 
+    textLinear: `Consider a hash table (size \${{0}}$) implemented with  "Linear Probing". 
+{{1}}
 {{2}}
-{{3}}
-{{4}}
     The hash-function is: 
-    {{5}} 
+    {{3}} 
     How is the state of the field after the operations?`,
     bottomTextLinear: `Please enter the state of the map after the operations in the format of an Array. You can keep free entries empty or just spaces.`,
     checkFormatLinear: "Please Provide an array only with numbers, commas and spaces",
   },
   de: {
     name: "Hashing",
-    description: "Korrektes Einfügen von Werten in eine Hashtabelle",
-    resize: "vergößerbare",
-    noResize: "nicht-vergrößerbare",
-    textLinear: `Consider having a {{0}} hash table (size of {{1}}) implemendted with  "Linear Probing". Initially the hash table is empty. Do the following operations:
-    {{2}} 
-    The hash-function is: 
-    {{3}} 
-    How is the state of the field after the operations?`,
+    description: "Füge den Wert korrekt in eine Hash-Map ein",
+    insert: "Wir **fügen** die folgenden Schlüssel ein: {{0}}",
+    delete: "und dann **löschen** wir die folgenden Schlüssel: {{0}}",
+    textLinear: `Betrachten Sie eine Hashtabelle (Größe \${{0}}$), die mit "Linear Probing" implementiert ist.
+{{1}}
+{{2}}
+    Die Hash-Funktion ist: 
+    {{3}}
+    Wie ist der Zustand des Feldes nach den Operationen?`,
+    bottomTextLinear: `Bitte geben Sie den Zustand der Karte nach den Operationen im Format eines Arrays ein. Sie können freie Einträge leer lassen oder einfach Leerzeichen verwenden.`,
+    checkFormatLinear: "Bitte geben Sie nur ein Array mit Zahlen, Kommas und Leerzeichen an",
   },
+}
+
+/**
+ * Generates a question with a hash map and operations
+ * @param random
+ * @param variant
+ * @param lang
+ */
+function generateQuestionBase(random: Random, variant: "linear", lang: "de" | "en") {
+  const tableSize = random.int(6, 8)
+  const tableSizeVariable = random.choice("mnxyzpvw".split(""))
+
+  // insert -> only inserting into hash table
+  // insert-delete -> inserting and deleting from hash table
+  const task: "insert" | "insert-delete" = random.weightedChoice([
+    ["insert", 0.85],
+    ["insert-delete", 0.15],
+  ])
+
+  const randomHashFunction = generateHashFunction(tableSizeVariable, tableSize, variant, random)()
+
+  const generated = generateOperationsHashMap({
+    random,
+    tableSize,
+    task,
+    mapStyle: variant,
+    hashFunction: randomHashFunction.hashFunction,
+  })
+
+  const hashMap = generated.hashMap
+  const insertions = generated.insertions
+  const deletions = generated.deletions
+
+  const insertTable = createTableForValues(insertions.map((i) => i.toString()))
+  const insertString = t(translations, lang, "insert", [insertTable])
+
+  const deleteTable = createTableForValues(deletions.map((i) => i.toString()))
+  const deleteString = deletions.length > 0 ? t(translations, lang, "delete", [deleteTable]) : ""
+  return { tableSize, tableSizeVariable, randomHashFunction, hashMap, insertString, deleteString }
 }
 
 export const hashingQuestion: QuestionGenerator = {
@@ -58,7 +93,7 @@ export const hashingQuestion: QuestionGenerator = {
     {
       type: "string",
       name: "variant",
-      allowedValues: ["linked", "linear"],
+      allowedValues: ["linear"], //Todo: Add linked
     },
   ],
 
@@ -84,49 +119,8 @@ export const hashingQuestion: QuestionGenerator = {
     let variant = parameters.variant as "linked" | "linear"
     variant = "linear"
 
-    const resize = random.choice([true, false])
-    const resizeString = t(translations, lang, resize ? "resize" : "noResize")
-
-    const tableSize = resize ? 2 : random.int(6, 8)
-    const tableSizeVariable = random.choice("mnxyzpvw".split(""))
-
-    // insert -> only inserting into hash table
-    // insert-delete -> inserting and deleting from hash table
-    const task: "insert" | "insert-delete" = random.weightedChoice([
-      ["insert", 0.85],
-      ["insert-delete", 0.15],
-    ])
-
-    const randomHashFunction = generateHashFunction(tableSizeVariable, tableSize, variant, random)()
-
-    const generated = generateOperationsHashMap({
-      random,
-      resize,
-      tableSize,
-      task,
-      mapStyle: variant,
-      hashFunction: randomHashFunction.hashFunction,
-    })
-
-    const hashMap = generated.hashMap
-    const insertions = generated.insertions
-    const deletions = generated.deletions
-
-    let insertTable = "\n|$" + insertions.join("$|$") + "$|\n|"
-    for (let i = 0; i < insertions.length; i++) {
-      insertTable += "---|"
-    }
-    insertTable += "\n|#div_my-5#|"
-    const insertString = t(translations, lang, "insert", [insertTable])
-
-    let deleteTable = "\n|$" + deletions.join("$|$") + "$|\n|"
-    for (let i = 0; i < deletions.length; i++) {
-      deleteTable += "---|"
-    }
-    deleteTable += "\n|#div_my-5#|"
-    const deleteString = deletions.length > 0 ? t(translations, lang, "delete", [deleteTable]) : ""
-
-    console.log(hashMap.keysList())
+    const { tableSize, tableSizeVariable, randomHashFunction, hashMap, insertString, deleteString } =
+      generateQuestionBase(random, variant, lang)
 
     const checkFormat: FreeTextFormatFunction = ({ text }) => {
       text = text.trim()
@@ -204,16 +198,12 @@ ${tableSeparator}
       return { correct: true, feedback: text }
     }
 
-    let question: FreeTextQuestion
-    // eslint-disable-next-line prefer-const
-    question = {
+    const question: FreeTextQuestion = {
       type: "FreeTextQuestion",
       name: t(translations, lang, "name"),
       path: permalink,
       text: t(translations, lang, "textLinear", [
-        resizeString,
         tableSizeVariable + "=" + tableSize.toString(),
-        resize ? t(translations, lang, "resizeFactor") : "",
         insertString,
         deleteString,
         randomHashFunction.hashFunctionString,
@@ -228,67 +218,4 @@ ${tableSeparator}
       question,
     }
   },
-}
-
-function generateOperationsHashMap({
-  random,
-  resize,
-  tableSize,
-  task,
-  mapStyle,
-  hashFunction,
-}: {
-  random: Random
-  resize: boolean
-  tableSize: number
-  task: "insert" | "insert-delete"
-  mapStyle: "linked" | "linear"
-  hashFunction: HashFunction | DoubleHashFunction
-}) {
-  let hashMap: MapLinked | MapLinProbing
-  if (mapStyle === "linear") {
-    hashMap = new MapLinProbing({ size: tableSize, hashFunction: hashFunction, resize: resize })
-  } else {
-    if (isHashFunction(hashFunction)) {
-      hashMap = new MapLinked(tableSize, hashFunction)
-    } else {
-      throw new Error("Provided function is not a valid HashFunction")
-    }
-  }
-
-  const insertions: number[] = []
-  const possibleValues = Array.from({ length: 25 }, (_, i) => i + 1)
-  for (let i = 0; i < (resize ? random.int(4, 5) : tableSize - random.int(0, 3)); i++) {
-    if (!resize) {
-      if (hashMap.getAmount === hashMap.getSize) {
-        break
-      }
-    }
-    const newValue = random.choice(possibleValues)
-    possibleValues.splice(possibleValues.indexOf(newValue), 1)
-    hashMap.insert(newValue, newValue.toString())
-    insertions.push(newValue)
-  }
-  const deletions: number[] = []
-  if (task === "insert-delete") {
-    for (let i = 0; i < random.int(1, 3); i++) {
-      const key = random.choice(hashMap.keys())
-      if (key !== null) {
-        hashMap.delete(key)
-        deletions.push(key)
-      }
-    }
-  }
-
-  return {
-    hashMap,
-    insertions,
-    deletions,
-  }
-}
-
-function isHashFunction(
-  functionToCheck: HashFunction | DoubleHashFunction,
-): functionToCheck is HashFunction {
-  return functionToCheck && typeof functionToCheck === "function" && functionToCheck.length === 2
 }
